Oh god... - Removed Warp Custom Bind (It was in the vanilla game, oops) - Transform control now follows rotation and moving by bind and by transform module - Transform control now works pretty well with Undo/Redo - Zoom now affects Transform control buttons scale - Removed some unnecessary binds, and reworked move/rotation by bind system a bit, so Custom Keybinds mod would work with GEAR - Swipe by shift from object now works fine with free move enabled - Swipe Teleportation bug is fixed (when you release shift while in select mode, there was some crazy teleportation. Also, it's a game bug, lol) - ImGui::InputInt in Edit Group Module is changed with ErGui::DragInt for consistency

+++ src/modules/EditGroupModule.hpp
@@ -28,10 +28,10 @@ namespace ErGui {
 	inline int maxZOrder;
 	inline int minZLayer;
 	inline int maxZLayer;
-	inline int minEnterChannel;
-	inline int maxEnterChannel;
-	inline int minMaterial;
-	inline int maxMaterial;
+	inline short minEnterChannel;
+	inline short maxEnterChannel;
+	inline short minMaterial;
+	inline short maxMaterial;
 
 	inline int minOrderValue;
 	inline int maxOrderValue;
